Remove ununsed installer code
