--- agent/installer/cli/main.go
@@ -1,12 +1,0 @@
-// Copyright 2021 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package main
-
-import (
-	"github.com/vmware-tanzu/cluster-api-provider-bringyourownhost/agent/installer"
-)
-
-func main() {
-	installer.Main()
-}
