--- agent/installer/doc.go
@@ -1,7 +1,0 @@
-// Copyright 2022 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Package installer contains the implementation of Kubernetes components installation
-// and uninstallation for a given OS and Kubernetes version.
-// All the components are packaged as an ImgPkg bundle hosted on OCI registries.
-package installer
