--- agent/installer/internal/algo/algo_suite_test.go
@@ -1,17 +1,0 @@
-// Copyright 2021 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// nolint: testpackage
-package algo
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestInstaller(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Installer Algo Suite")
-}
