--- agent/installer/internal/algo/doc.go
@@ -1,6 +1,0 @@
-// Copyright 2022 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-// Package algo contains the interal implementation of the actual steps
-// required for installation and uninstallation of Kubernetes components.
-package algo
