+++ installer/bundle_builder/Dockerfile
