+++ installer/bundle_downloader.go
