+++ installer/bundle_downloader_test.go
