+++ installer/installer_suite_test.go
@@ -1,8 +1,7 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// nolint: testpackage
-package installer
+package installer_test
 
 import (
 	"testing"
@@ -11,7 +10,7 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-func TestInstaller(t *testing.T) {
+func TestCloudinit(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Installer Suite")
 }
