+++ installer/registry.go
@@ -86,22 +86,14 @@ func (r *registry) ListK8s(osBundleHost string) []string {
 	}
 
 	// os host
-	for k8s := range r.osk8sInstallerMap[r.resolveOsToOsBundle(osBundleHost)] {
+	for k8s := range r.osk8sInstallerMap[r.ResolveOsToOsBundle(osBundleHost)] {
 		result = append(result, k8s)
 	}
 
 	return result
 }
 
-// GetInstaller returns the bundle installer for the given os and k8s version
-func (r *registry) GetInstaller(osHost, k8sVer string) (osk8si osk8sInstaller, osBundle string) {
-	osBundle = r.resolveOsToOsBundle(osHost)
-	k8sBundle := r.resolveK8sToK8sBundle(k8sVer)
-	osk8si = r.osk8sInstallerMap[osBundle][k8sBundle]
-	return
-}
-
-func (r *registry) resolveOsToOsBundle(os string) string {
+func (r *registry) ResolveOsToOsBundle(os string) string {
 	for _, fbp := range r.filterOSBundleList {
 		matched, _ := regexp.MatchString(fbp.osFilter, os)
 		if matched {
@@ -112,13 +104,41 @@ func (r *registry) resolveOsToOsBundle(os string) string {
 	return ""
 }
 
-func (r *registry) resolveK8sToK8sBundle(k8s string) string {
-	for _, k8sBundle := range r.filterK8sBundleList {
-		matched, _ := regexp.MatchString(k8sBundle.k8sFilter, k8s)
-		if matched {
-			return k8sBundle.k8sFilter
-		}
+// GetSupportedRegistry returns a registry with installers for the supported OS and K8s
+func GetSupportedRegistry() registry {
+	reg := newRegistry()
+
+	var empty interface{}
+
+	{
+		// Ubuntu
+
+		// BYOH Bundle Repository. Associate bundle with installer
+		linuxDistro := "Ubuntu_20.04.1_x86-64"
+		reg.AddBundleInstaller(linuxDistro, "v1.21.*", empty)
+		reg.AddBundleInstaller(linuxDistro, "v1.22.*", empty)
+		reg.AddBundleInstaller(linuxDistro, "v1.23.*", empty)
+
+		/*
+		 * PLACEHOLDER - ADD MORE K8S VERSIONS HERE
+		 */
+
+		// Match any patch version of the specified Major & Minor K8s version
+		reg.AddK8sFilter("v1.21.*")
+		reg.AddK8sFilter("v1.22.*")
+		reg.AddK8sFilter("v1.23.*")
+
+		// Match concrete os version to repository os version
+		reg.AddOsFilter("Ubuntu_20.04.*_x86-64", linuxDistro)
+
+		/*
+		 * PLACEHOLDER - POINT MORE DISTRO VERSIONS
+		 */
 	}
 
-	return ""
+	/*
+	 * PLACEHOLDER - ADD MORE OS HERE
+	 */
+
+	return reg
 }
